requirements.txt
matplotlib==3.7.2
mplsoccer==1.1.10
numpy==1.26.3
pandas==2.0.3
pydantic==2.7.2
highlight_text
requests==2.28.1
adjustText==1.2.0
python-dotenv==0.21.1